% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontlevel_to_foldchange_and_ngenes_plot.R
\name{ontlevel_to_foldchange_and_ngenes_plot}
\alias{ontlevel_to_foldchange_and_ngenes_plot}
\title{Ontology level to fold change & ontology level to n genes facet plot}
\usage{
ontlevel_to_foldchange_and_ngenes_plot(
  all_results_merged,
  hpo,
  phenotype_to_genes
)
}
\arguments{
\item{all_results_merged}{HPO EWCE results <data.frame>}

\item{hpo}{The HPO data object from ontologyIndex}

\item{phenotype_to_genes}{phenotype gene list annotations from hpo
phenotype_to_genes.txt}
}
\description{
This creates two plots faceted which show the relationship between ontology
level and fold change in specific expression and number of associated genes
in the rare disease EWCE results and HPO. The phenotype name column is called
list.
}
